getch_l.c: improve restore_getch for general usage
The function restore_getch can't be used to also restore the non-
blocking state. It also doesn't check if the tty was actually
changed.

So make it do so to be able to use it also within the code to
restore the tty (e.g. for scanf stuff) and to prepare it again.