ugtrain: add adaption script execution
Now we can run the following command:
$ ugtrain -A chromium-bsu64

The configured adaption script is run from a relative path which is
always relative to the config file path.

The line in the config file could look like this:
adapt_script adapt/chromium-bsu_adapt64.sh

For security reasons this will fail as root.