ugtrain: add stack offset discovery
We send the discovery config to the discovery lib and parse the
output from the FIFO.

We don't call the actual trainer code but instead run in a separate
loop to read the stack offset from the FIFO. We use ptrace as well
to detect if the game is still running. The new stack offset isn't
written back to the in-memory config, yet.