libmemdisc: determine heap start
We've noticed that the heap address filter became useless with heap
randomization activated in the kernel.
We can do against this the same thing as what we do against stack
randomization: Using offsets. But we start with determining the
heap start. Using a malloc before everything else does the trick.