ugtrain: introduce 'adapt_required' config parameter
If we use dynamic memory and the following option is set in the config:
adapt_required 1

Then, we ask if we should run the adaption script and the discovery.
If yes, then it works like the regular adaption. We don't ask the user
if he explicitly requests the adaption by CLI parameters.