ugtrain: discovery: enhance stage 3 preparation
We need at least 3 arguments in stage 3: Heap start and end offset
and the memory allocation size (discovered in stage 1 or 2).

So the minimum discovery string for stage 3 looks like this:
'3;0x0;0x0;<size>'

For convenience we determine the code start and end address
automatically now with a long shell command. It is based on
'objdump -p'.