examples: warzone2100: use value names from game source
We've earned the right to look into the game source to check how
our found static values are really called. Interesting is that the
power is stored in an array for all players and that it is a fixed
point value. We were lucky to find it as i32 as it is an i64 in
fact. The lower 32 bit are used for the part after the point. But
that part doesn't matter for ugtrain.