ugtrain: support multiple memory objects per class
In Chromium B.S.U. it is easy that we only have a single
HeroAircraft object. But in Warzone 2100 we have 47 Droid objects
which are created from the same malloc call upon campaign start.

So we need to support multiple memory addresses per dynamic memory
config in ugtrain. In order to reduce the output, we only print the
values in first memory object per class continuously.