ugtrain: add dumping of all mem. obj by '>' key
Now, we only need to press the '>' key and the memory objects of
all classes are dumped into files like this:
0_000.dump

This way we can analyze the memory contents better. Before the
underscore is the class ID and after it is the object ID within
the class. Existing files are overwritten. These are binary files
but it is possible to read and convert them with a HEX tool like
'xxd'.