add 'tools' directory with 'ugsimfind'
ugsimfind is very useful if you've got a memory class in which also
opponents are stored like the Droid class in Warzone 2100. You need
a value which is the same for all of your own units but differs for
opponents so that you can detect which units are your own ones.

ugsimfind finds similarities in binary memory object dump files on
byte basis. It tells you for a file address the byte value in hex
and as char, its count and all files in which this value has been
found at that file address.