libmemhack: support freeing of multiple mem. obj.
We use the free space in the page-sized config buffer at the end
to store the memory addresses returned by malloc. We can keep
track of upto 504 memory addresses this way. As all frees have to
be compared with this list, we detect the end of the list earlier
with a magic invalid pointer value which is not NULL and fill the
whole array of memory addresses with that value initially.
The last memory address is always an invalid pointer. This saves
us one compare per loop cycle in 'free'.