cfgparser: prepare dynamic wish value handling
Wouldn't it be cool to let ugtrain detect the wish value
dynamically for us? We just tell it maximize or minimize it and
we start at 0. This can help to find the limits in which a value
can be stored. We start with the parsing support of the wish values
"max" and "min".

This also fixes the not a number check in parse_value(). In the
float '.' check the brace is at the wrong location so that
arbitrary chars can be set. But now we need that check working
correctly.