ugtrain: add wish value from mem. address support
We've noticed issues when setting a value higher than 100% for
convenience. E.g. Warzone 2100 3.1.0 has an assert for the droid
body points telling us that we've set it too high. Those assert
failed messages are nasty and are slowing down the game. So we
have to support dynamic wish values from a memory address. The
ugsimfind can help locating that memory address. Then, we get
the maximum value for each object from within the object.

We detect an address as value with the prefix "0x".