discovery: fix stack offset write back in stage 5
The handling of multiple stack offsets introduced a bug in the
dispaying and writing back in the single object single stack
offset adaption in postproc_stage5(). We are writing back the
memory address of our stack_offs[] array. So use stack_offs[0]
instead.

For now we don't support adation of more than a single stack
offset but that support will follow.