ugtrain: add support for dynamic memory object checks
We introduce a new reserved config entry name "checko" which works
basically like a regular check but leads to removing the currently
active dynamic memory object from the vector of memory addresses to
be cheated if the check fails.

This helps us e.g. in Warzone 2100 to only keep track of our own
droids.

In the hot path (when attached to the game memory) we only set the
memory address of the object to NULL. After detaching, we can
remove these marked addresses from the memory address vectors of
our dynamic memory entries and print information about doing so.