ugtrain: add watch value support
We can use the dynamic value "watch" now and the value from a
memory address is just watched but not changed. Parsing ends then
after the value. As an alternative it is possible to use a regular
value and use 'w' as activation state instead of 'a'. Then, the
the watching can be toggled by key bindings: Watch/unwatch. But
watching is directly active upon game start then.