ugtrain: introduce 'dynmemign' config command
With that we get a method to configure a stack offset which is used
for adaption but not for production. This helps the adaption to
know that there might be unexpected stack offsets.

The plan is to sort the stack offsets and to ignore e.g. always
the highest stack offsets. If these stack offsets couldn't be
discovered, it doesn't matter for us.