linuxhooking: unify use of 'common.h', increase config space
We use all the helpers and types from common.h, now. This helps
avoiding coding stuff twice. Forthermore, we increase the max.
number of stack offsets from 4 to 8 and we use 4 memory pages
instead of only one for the config in libmemhack. This is
important due to the fact that the game can become unstable
if we don't keep track of enough memory addresses.