ugtrain: introduce reverse message parsing for discovery
We've noticed an issue in discovery stages 3 and 4 when there is
more than a single match. The symbol lookup is still quite
inefficiant and slow as we read symbols uncached one by one from
the disassembly.

What matters for us is (one of) the last matching malloc call(s)
as it is very likely kept until game end. When reversing the
search for the matching malloc call, we can interrupt the symbol
lookup directly after the relevant malloc. This is why we do it
this way, now.