discovery: add DEBUG_PARSING compiling option
When debugging the file read and the parsing, we don't want the
input file to be created from scratch over and over again. So
we add a compiling option to write the output from the FIFO to
/dev/null in the stages 1, 2, 3 and 4 loop if activated.

It is deactivated by default. So regular program execution is
not affected. It is used for local debugging and testing only.