examples: add adaption to Warzone 2100 64-bit
The adaption is not completely ready yet as there are also major
differences between game versions but for version 3.1.0 it works
at least. There are also still issues with the multiple stack
offsets.

We also increase the maximum body points for Droids to 2000.
From these the actual body points are read. This makes our
droids more difficult to destroy.