remove scanmem_extension directory
If upstream scanmem doesn't want to have an array/vector dump, then
why not write an own tool to do just that?

That sounds much better than forking and hacking scanmem. They are
plain not aware of dynamic memory objects which are put to heap
memory one after another of the same memory class and therefore
size. Anyway, ugtrain has a dump of the dynamic memory objects it
captures by now ('>' key). So we can dump and analyze them this
way much better. A tool like this would only be needed for Wine
as we can't hook malloc() and free() there, yet.