discovery: discover dynmemign optionally in stage 5
This is a preparation for a smarter stage 5 adaption. The idea
is that the stack offset in dynmemign doesn't have to be found.
We will later sort the given and the discovered stack offsets
so that they are put to the correct locations no matter of the
order they have been discovered.