discovery: enhance stage 5 by sorting stack offsets
With sorted stack offsets it doesn't matter anymore in which order
they have been triggered within the game. The order of the stack
offsets remains often the same. So we have very good chances to
hit successful discovery and adaption with that. But with more
than one stack offset a risk remains. So we better ask the user
if the automatic adaption really did the right thing per dynamic
memory class.