libmemdisc/hack: use frame pointer, not stack pointer
We don't want to search in our own stack usage for code pointers.
This is why we need to get the start of our current stack frame.
From the GNU backtrace() source we know that this is possible with
'__builtin_frame_address (0)' in an arch independent manner. Now,
we don't have any arch specific stuff in the C/C++ code anymore.