ugtrain: add '--glc' option for video recording
We want to run the game in our own process group so that we don't
need root permissions for ptrace. So if we want to record our
cheating, we need to run the game with 'glc-capture' instead of
'ugpreload'. We give over the user-specified arguments to it and
we're done. This also works for pure static memory configs. With
'--pre-cmd' we can specify arbitrary preloader commands.

Note: glc-capture must support the custom '--preload' option for
      dynamic memory cheating as it has to preload libmemhack.