system: do major enhancements for scanmem
We want to run scanmem and the game in the same process group. This
way we don't require root priviledges for it. So we introduce the
new function run_pgrp_bg() for that. It is highly specific and
hacky. It is possible that the game becomes an orphan as we can't
wait for it (due to the exec call for scanmem in the parent).

We need a method to catch orphans (wait for them to exit). The idea
is to use proc_to_pid() with a delay in a loop. So we have to move
the error message there into ugtrain's main() function.

As a further preparation we replace fork_wait_kill() with
fork_proc() and use the functions wait_proc() and kill_proc() for
having the same functionality. This way we can place our orphan
catcher to wait for the orphaned game process.