ugtrain: Introduce scanmem support
We can use scanmem as regular user, now. It is usable in the
discovery stages 0..4 and with pure static memory together with
'-P'. We catch the game process if it becomes an orphan and we use
the chance here to improve discovery and adaption logic as well.

Known issues:

Scanmem reads the available memory regions only once upon start
and only searches in these. So the first time the intersting
object is visible, the scanmem command 'reset' has to be entered.