libmemhack: double the config size for 64-bit
It is unfair that we can handle double the number of memory objects
on 32-bit systems. So let us use 'sizeof(long)' to set the number
of config pages. It is almost the same number of pointer compares
this way on 32-bit and on 64-bit systems and almost the same number
of max. memory objects per class.