examples: warzone2100: 32-bit: update checks, add heap ptr
We've noticed that it is required to put the object check to each
value we can control via keyboard as the first value in the class
could be deactivated.

We've also found a heap pointer in the hex dump of a Structure
which is only a little bit higher than the Structure pointer. It
is likely that this one is later allocated. So it could be possible
that further interesting values are hiding in that memory object.
We start with watching that pointer.