libmemdisc: introduce calloc() hooking
Hooking calloc() is really special as dlsym() uses calloc. So for
the first call to calloc we have to return static memory. But don't
worry: There will be no free() call for it.

Details:
http://www.slideshare.net/tetsu.koba/tips-of-malloc-free

We move the malloc post-processing into a separate inline function.
We use it for calloc() as well. This way we don't have to duplicate
code from the malloc() function. The result is that we see all
callocs as mallocs in the discovery output.