ugtrain: enhance checks, improve cfg output
We introduce CHECK_EQ (==, 'e') and use the chance to rework the
check processing. If a "check" entry has no 'g', 'l' or 'e', then
parsing should already fail. We can detect that, now. So use it.

Also the config output is very difficult to read, especially with
a lot of checks and classes. So introduce some newlines and spaces
for indention.