examples: warzone2100: 64-bit: research and build droids faster
Do the same thing as already done for 32-bit.

In the 32-bit config there is a subject pointer watcher used as u64.
Fix it by using u32.

In 3.1.0 there is no "powerAccrued" for factory or research facility
any more. They use build points remaining to build instead of time
remaining to build.