ugtrain: introduce 'game_call' support
There are games which are called by a shell script. So process name
and game call differ from each other. With this support static
memory cheating is still possible. But adaption and dynamic memory
cheating won't work this way. For that the direct binary call is
required.

Example: 'sauerbraten' calls 'sauer_client'