linuxhooking: only care for the game process only once
There are games which are called from a shell script or we use
a tool like glc-capture. But we don't want these to activate the
hooking. If everything is active already, there is no need to read
from the already empty input FIFO. The expected game process name
is transfered via an environment variable.