ugtrain: introduce 'game_binpath' support
There are games which are called by a shell script. So process name
and game call differ from each other. Also the regular 'game_path'
is unsuitable for disassembly as it relates to the game call. This
is why we need a separate path to the game binary. If it's not set,
the 'game_path' is assumed.

Example: 'sauerbraten' calls 'sauer_client'