ugtrain,ugpreload: integrate 'ugpreload' into 'ugtrain'
This introduces the following new files:
* src/preload.h
* src/preload_l.c

And this removes the ugpreload source. There is no need to having a
separate preloader as ugtrain can set LD_PRELOAD and the env vars
for libmemhack/disc. This removes a lot of duplicate code as we can
call run_game() with that. It doesn't really make sense to call
ugtrain as root if that can be avoided.