ugtrain: remove 'ugpreload' occurences and update doc
This removes the src/linuxhooking/examples directory with all shell
scripts inside. The manual method is way too complex. This is why
we've integrated that into ugtrain. So there is no need for it
anymore.

In the README, in the help text and in the dynamic memory
documentation there are still ugpreload mentions. So update them
for the new behavior.