ugtrain: fix some data types
We want to use i32 instead of int to show that this is a signed
integer of 32 bit width. It becomes more important when it comes
to i8 vs char where we want to differentiate between integer use
and char/string use.
Only the getch handling is an exception as it could be used for
other software tools as well without common.h depend this way.