options: optimize and improve assumptions
We have a function use_libmemhack(). So let's introduce
use_libmemdisc() as well and make them inline.
It is a bit unclear which assumptions handle what. So add some
comments. Currently we only handle adaption and preloading/starting
the game. Optimize the checks for that and if we have a preloader
command like "--glc" set, then also run the game.