ugtrain: never use a shell to run the game
We are lazy and let a shell convert our command string into a
command vector for execv() when a preload command or game options
are given. But double code is bad. So always put a command string
together, split it into a string vector and convert that one into
a C string vector. Sounds cumbersome but should be the most
flexible way. This should also improve security and performance.