ugtrain: handle exec errors and rework wait handling
If exec fails, then we should handle that accordingly by exiting.
Also the wait handling is not correct. We can't wait for an already
running game process. So maintain a call pid which is set after
exec and use that one for wait(). If it is < 0 and the game pid
is found, use the other wait method.