ugtrain: do cleanup, remove 'NONE' cfg handling
Some variable and function names are still a bit fuzzy. So rename
them to something more obvious. Change some comments, split some
lines and remove the 'NONE' cfg handling. We always have to test
and special treat the 'NONE' handling and we can't ask the user
everything. So it really makes sense to write a config, even if
there's only the process name inside.