system: introduce sleep_sec_unless_input()
Finally we've found the secret of the Linux command 'top' how to
only sleep unless there is input. The libc select() function is
used as the kernel notifies on the stdin file descriptor. Ugtrain
becomes much better responsive to user input with that and still
doesn't risk activation list corruption.

We also use the chance to move the other inline functions to the
system header and to improve some minor things.