discovery: do clean-up
In the meanwhile we know what we need and what not for successful
discovery. There is no need for looking at free() calls at all. We
should keep the option to record them but using them for matching
makes no sense. So remove that call back function and also remove
commented out debugging output. If we need to debug something, we
can work with a debug define.