discovery: optimize symbol lookup in stages 3 and 4
At the moment we run 'objdump -D' for each and every symbol. With
a small game binary like 'chromium-bsu' this is okay more or less
but with bigger binaries it is too slow.

So disassemble only once and write the output to the following file:
/tmp/memhack_disasm

The symbol lookups just read that file and search in it for the
symbol at the related jump-back code address. The disassembly file
is removed afterwards. For that also some system function changes
are required.