system: improve use_shell handling and run_pgrp_bg()
If the command vector cmdv is set to NULL, then it is already known
that the shell has to be used. So we can remove that parameter.

In run_pgrp_bg() the use_shell param is used to determine the
method to get the game PID. But using the PID of the game call
instead of the game process is unsafe. So always determine and use
the game process PID.