linuxhooking/ugtrain: introduce PIE support
Some Linux distributions use the GCC options -pie and -fPIE
by default for hardened security. Together with ASLR, the binary
is loaded each execution to another memory location. So we have to
find out that memory location and use it as an offset. This
applies to code addresses on the stack and static memory as well.

The kernel knows about the location in /proc/$pid/maps. As we have
to fork in order to read from that file, we also have to remove
libmemdisc/hack from LD_PRELOAD. This also helps us hiding from the
game process that libmemdisc/hack is preloaded. ;-)

Finding this out also involved a lot of debugging which can be
simplified and unified between libmemdisc and libmemhack by
introducing pr_dbg(), pr_out() and pr_err() in new files
libcommon.[ch]. We always need the prefix to show that this is
neither game output nor ugtrain output. So we do it similar to the
Linux kernel and define some print helpers.

pr_dbg() is printf(PFX ...) and is not compiled by default.
pr_out() is fprintf(stdout, PFX ...).
pr_err() is fprintf(stderr, PFX ...).