ugtrain: rework config entry type and value handling
Current state is that we use the three type values is_signed,
is_float and size separately for value entries and check entries
in the config. The values are used as an i64 value which is
casted into the real data type. This is quite complex and can be
improved for better performance as well as readability.

So put the type values into a separate structure and use it as a
function parameter whenever different behavior depending on the
type is required. Also replace i64 values with an 8 byte union
'value_t' (scanmem uses the same). This avoids a lot of casting
within the hot path.