ugtrain,discovery: sleep 1 sec unless FIFO input
Especially, in stages 1, 2, 3 and 4 there is an infinite loop
without a sleep causing a CPU core to reach 100%. Use a sleep
unless input on the input FIFO to fix this.

In the main loop of ugtrain we already react on user input but
there it also makes sense to react on the input FIFO. So do it
and use a second file descriptor as argument. If it isn't needed,
use -1. We don't use va_list to keep this inline.