cfgparser: add config option to disable preloading
On the MeeGo 1.2 Harmattan platform the regular user is not allowed
to use LD_PRELOAD. So the idea is to patch the ELF binaries with
patchelf to inject the library libmemdisc or libmemhack. In this
case the command line option '-P -' or the config option 'no_preload'
is required. The behaviour is the same but LD_PRELOAD isn't set.