ugtrain: introduce memory maps handling
Doing text search in /proc/$pid/maps isn't the best method when
thinking about PIC and PIE. So parse it the same way as scanmem
does it. Send the code offset synchronized to libmemdisc and
libmemhack as well to implement PIE support again. The libraries
only adapt their configs to PIE.
Ugtrain converts the addresses back to the addresses within the
executable.