examples: add 64 bit sauerbraten config + adaption
This adds the 64 bit Cube 2: Sauerbraten config with adaption. It
is quite similar to the chromium-bsu config as there is also a
single object to cheat. But special is that game call and game
binary differ as sauerbraten is a shell script which executes
sauer_client.

Health, armor and ammo are frozen.

This even works in multi-player mode but you are not invulnerable
there as the current health and the done damage are sent to the
client in the same network package.