examples: chromium-bsu: simplify the config
Do it like for Cube 2: Sauerbraten that ammo values start at 0 as
also the ammo array starts at 0. Furthermore, we don't need to
control each weapon one by one. It is enough having one key for all
ammo, one for shields and damage and one for the lives. Others
still can configure it differently of cause.