adaption: introduce malloc size adaption
With Cube 2: Sauerbraten we have noticed that there are cases in
which it is required to adapt the malloc size as well. But this is
only safe if the configured offsets within the objects don't
change. Also the adaption result format changes. This is why this
patch also changes all adaption scripts.