ugtrain: add PIE detection on ARM
With the armv7l in the Nokia N9 MeeGo 1.2 Harmattan smart phone the
static load address is at 0x8000. Use it to correctly detect that
this smart phone doesn't use PIE. Also clarify the comments in the
source code for this.

TODO: Read the ELF header to detect PIE in an universal way.