linuxhooking: introduce libprivacy
The ugtrain is intended for single-player FOSS gaming only. With
libprivacy preloaded it is possible to enforce that for not
annoying others by mistake or to send crash reports to game
developers which were caused by ugtrain. If abuse of ugtrain
increases, we can preload libprivacy by default.

For now just block all calls to socket() and socketpair() which
don't use AF_LOCAL to make it simple. This has been tested against
the examples chromium-bsu, warzone2100 and sauerbraten.