system_l: get the most recently created pid
At the moment we get the first created pid. This turned out to be
bad if a game process became a zombie or is disfunctional in any
other way. It was embarrassing hitting this issue during the
presentation of scanmem and ugtrain on Berlin Hack and Tell #30.
So replace "pidof -s $proc_name" with something like "pidof
$proc_name | cut -d ' ' -f 1" to get the most recently created pid.

This also allows us to run multiple game instances with cheating by
ugtrain.