ugtrain: rework running the game
Also with pure static memory it should be possible to run multiple
game instances and attaching to the most recently started one. So
rework that. With dynamic memory it also doesn't make much sense
not to start the game ourselves. It is too complicated, insecure
and we don't provide a separate preloader anymore. So fail there
without '-P' given. This also helps simplifying documentation.