system: rework the process checks
At the moment we don't return the actual state from the process
check but instead only if we expect the game process to be running
or not. That's bad for debugging. Also maintaining two methods is
too complicated. Furthermore, on the MeeGo 1.2 Harmattan platform
we've noticed a loader which ends very late (while the game process
is already running for a second). We've attached to the loader
instead of the game.

So rename the function pid_is_running() to check_process(), remove
the waitpid() method and return one of the following states:
PROC_RUNNING, PROC_ERR, PROC_DEAD, PROC_WRONG or PROC_ZOMBIE.
An error in the detection is expected to be the same like a running
process. Check the called process periodically and try to get the PID
of the game process as a loader may have forked the game. This should
be the safest method. We give the detection upto 5 seconds before the
game process should be running for sure.