debian: rename debian/ to debian_harmattan/
At the moment we maintain two git branches 'master' and 'harmattan'.
The harmattan branch always has to be rebased to master as it just
has the debian packaging for MeeGo 1.2 Harmattan on top. This is
cumbersome and we still want to show that we don't have any Debian
packaging for regular desktop PCs.

So rename the debian directory to debian_harmattan. A further patch
will enhance autogen.sh to set a 'debian' symlink to it if the MeeGo
1.2 Harmattan platform is detected.