cfgparser: add support for growing memory objects
We've found a vector of pointers in a game which grows by calling
realloc(). It grows by the size of a pointer each time until it
reaches its final size. The allocation of the first pointer is
always a malloc(). With the current code we would have to copy
the same dynamic memory config over and over for all possible
sizes. This is cumbersome and looks like this:

dynmemstart GrowingClass 16 0x754 0x8
dynmemend
dynmemstart GrowingClass 24 0x754 0x8
dynmemend

So introduce growing object support in config parsing. Introduce
a config command "dynmemgrow" which is used within the dynamic
memory config. Make it possible to use it instead of "dynmemstart"
as well in case the initial malloc() is unknown. Support only
growing by adding for now and introduce a vector for storing the
current size for each allocated object of this growing class.
Looks like this:

# Initial malloc() and growing by realloc() known

dynmemstart GrowingClass 8 0x73c 0x8
# dynmemgrow <size_min> <size_max> <operator><growing size> ...
# ... <code_addr> <stack_offs>
dynmemgrow 16 40 +8 0x754 0x8
dynmemend

# Only growing by realloc() known, malloc() config becomes invalid

# dynmemgrow <class_name> <size_min> <size_max> ...
# ... <operator><growing size> <code_addr> <stack_offs>
dynmemgrow GrowingClass 16 40 +8 0x754 0x8
dynmemend