memmgmt: rework output_dynmem_changes()
If there is a memory object created and deleted in the same moment
in time, then it looks like this:

===> Obj. MagicClass created 1 time(s); now: 0
===> Obj. MagicClass freed 1 time(s); remaining: 0

It can be confusing that the creation shows that there is no memory
object at the moment. Furthermore, output_dynmem_changes() uses a
lot of 'cfg_en->dynmem' which can be simplified.

So assign 'dynmem = cfg_en->dynmem;' and use 'dynmem' to simplify
code. Also remove the '; now: 0' output if objects have been freed
for that class as well. Looks like this:

===> Obj. MagicClass created 1 time(s)
===> Obj. MagicClass freed 1 time(s); remaining: 0