autotools: fix build with glib and multilib enabled
On 64 bit systems with the configure options --enable-multilib and
--enable-glib set we have the problem that the glib check only
detects if the 64 bit glib is available but not if a 32 bit glib
is available as well. So disable the 32 bit glib build in that
case. The result is that there are 32 bit libraries hooking the
standard C memory allocation functions and 64 bit libraries hooking
glib memory allocation functions like g_malloc() as well.