ugtrain: read again from FIFO while being attached
Especially on the MeeGo 1.2 Harmattan platform with only limited
memory resources there is the issue that the game can't be started
from time to time or crashes during gameplay when there is a free().
The game trainer attaches to temporary memory objects and when they
get freed, there is a ptrace() R/W access to a heap address which is
not located within a valid heap region. This causes the game process
to crash with SIGSEGV. It is a race condition as it only happens if
the free() call is triggered after FIFO parsing before attaching to
the game memory with ptrace(). While being attached, there must be a
free() in the memhack FIFO as the game process is completely frozen
and as we write the free() call to the FIFO before executing the
actual free().

So read again from the input FIFO while being attached. Mallocs and
frees get processed as we can't separate them. This also requires
changes to the dynamic memory allocation/freeing output. The thing
is that we need to differentiate between free() calls and kicked
out objects there.

Fixes: GitHub issue #3