libmemdisc: add PIE support for pointer to heap obj. discovery
While testing all discovery features again it has been noticed that
the rarely used pointer to heap object discovery doesn't work
on systems with PIE on everything.
So add the PIE support by sending one code offset for the pointer
to heap discovery code address or the pointer offset and one for
the discovery stage >= 3 to libmemdisc.