libmemdisc: remove memdisc_exit() as it segfaults
We've noticed that there is always a segfault in libc when ending
the game. The library destructor was just another try to finally
flush the output buffer. It is ineffective as the _exit() function
closes open file descriptors before calling the library destructor.
We use atexit() to register our flushing function to exit() now.

So remove memdisc_exit() completely. It doesn't hurt us to keep
file descriptors opened until the game exits.