libmemdisc: introduce another FIFO for discovery output
We've noticed the need to react immediately to changes in the game
process during discovery. But in stages 1, 2, 3 and 4 everything is
written to a file. So introduce a new FIFO "memdisc_out" for that
and use "memhack_out" the way it's meant to be: for immediate
actions without caching. Convert the PIE handling to use it.

Following commits will use immediate actions in case of library
load so that the /proc/$pid/maps file can be re-read.