libmemdisc: hook dlopen() for PIC support
As libmemdisc is loaded as the very first library, we need to
inform ugtrain about the libraries loaded later on. With the
library name it is possible to read from /proc/$pid/maps again to
get the load address of the library and to handle PIC correctly
during discovery stage 3/4 postprocessing.

The output written to the "memhack_out" FIFO looks like this:
l;libGL.so.1

Following commits will read this data from the FIFO.